Added a symbol for 'external source documents'.
I've spent A LOT OF TIME TAKING INTO CONSIDERATION EVERY HTML5 SYMBOL IN
EXISTENCE, so this one is as good as it gets.
For now, I just plan to let doc.info point to a (local) document id and then
set href to _show/edit/<id>, that is, the futon page. The document should be
filled with attachments and annotations lining out the source of the features
that are displayed in the map page.
This pretty much voids my responsibility (yay!!) to document every possible
detail of the uploaded features: I just refer everyone to the source documents
and then say you can add whatever you find there.
I foresee the necessity to use an extra document because:
- it should be hidden from all indexes (by using no type),
- it will be referred to by a bunch of features.