Reviewed #filter [name=content]'s autocomplete function.
This seems more like it: the source callback moves through the view results but
only as long as it needs to. The view results are cached anyway, so I noticed
just a small latency.
As a bonus, I can use the cached view results for the content pre-filter, which
would void the confusing "<combined results>" autocomplete suggestion. Nice!!!