Clustered query to _view/field-type-value for autocompletion.
After all that is what the whimsical hierarchy is for. Common use-cases are
having all or just one type selected when searching for a certain field
content. Both cases now result in one GET.
As a collatoral, numbers are included in the suggestions. I forgot why I've
ruled these out in the first place though.