Improved sequential content pre-filter query.
The sequential download is meant to be identical to the autocomplete source
handler:
- 100 rows with group=true,
- then chunks of 2000 rows with reduce=false if any.
Therefore if the user types in a guess for his required field, most of the rows
are already download into the browser cache. The search result appears to pop
up instantaneously while in reality it isn't!!! IN-GENIOUS!!!!
It's the best I've managed to come up with since I sadly had to remove
value.doc.content to protect the view against reduce overflow errors. I should
add some comments explaining the logic and whatnot.