Added licensing section to readme and annotations.
I am unsure if this project couldn't be a gold mine some day, so I'll restrict
access to it as much as possible. Plus some of it is well written and will be
considered Gold Standard soon, so no one must be allowed to just copy it into
his own crappy project without me ever knowing.