Split indexing process in chunks for smoother page load.
Sorting that big stupid document array freezes the page for three seconds on my
Lenovo X60. And I don't even need the index right away.
So I split the whole sort process into nice and easy chunks of 500 keys. Each
chunk is sorted and merged into an accumulated list by insertion. The page
continues to load and react, and the index will be available after a couple of
seconds.