Replaced the workaround for Chrome with a timeout.
My best guess at what the problem was would be that CouchDB sent something like
an invalid HTTP response, to which FF reacted by closing the connection, and
Chrome by staying connected forever.
Since the request is for a limited amount of documents and geometry, it should
in any case take less than three seconds. A timeout is the least confusing way
to manage this.