Removed POST _list/bbox-filter for installed GeoCouches.
Because for limited bandwidth or any reasonable DSL connection, the stupid POST
with 8000 or so keys takes waaay too long. _spatial/bbox on the other hand is
only 300kB gzipped.
The remaining problem is the huge amount of CPU time it takes to parse the
response and hash-list ids. It causes the map to stall a little.