Patched export w/ compressed indexes AGAIN.
Because apparently I am not any good with the computers and also I don't like
testing because it sounds unpleasant.
Meh anyway it seems I started with an empty array of documents and tried to get
documents[-1].toString() with it. That of course broke it, regardless of the
out-of-bounds check a couple of lines below.
So I put in an extra lower bound, removed said out-of-bounds-check and
corrected for the offset.