Improved elimination of duplicate documents w/spreadsheets.
That part was removed from _list/export and re-introduced to the map page.
I have a bunch of documents with each attached to some 100 geometries. It took
waaaay too long to download a spreadsheet of those documents, because the url
consisted of 1000 keys instead of just a bunch, as it should be.